consolidate hard reset, SVID discovery, and discover identity traits into main PD trait
