+++ type-c-interface/src/controller/mod.rs
@@ -3,14 +3,11 @@
 use embedded_services::named::Named;
 use embedded_usb_pd::PdError;
 
-pub mod discover_identity;
 pub mod electrical_disconnect;
-pub mod hard_reset;
 pub mod max_sink_voltage;
 pub mod pd;
 pub mod power;
 pub mod retimer;
-pub mod svid;
 pub mod type_c;
 
 /// Controller ID
