+++ type-c-interface/src/port/mod.rs
@@ -1,11 +1,8 @@
 //! Type-C port related code
-pub mod discover_identity;
 pub mod electrical_disconnect;
 pub mod event;
-pub mod hard_reset;
 pub mod max_sink_voltage;
 pub mod pd;
 pub mod power;
 pub mod retimer;
-pub mod svid;
 pub mod type_c;
