+++ type-c-service/src/driver/tps6699x.rs
@@ -721,6 +721,76 @@ impl<M: RawMutex, B: I2c> Pd for Tps6699x<'_, M, B> {
 
         { self.tps6699x.lock_inner().await.set_tbt_config(port, config_reg).await }.map_err(|e| self.log_error(e))
     }
+
+    async fn hard_reset(&mut self, port: LocalPortId) -> Result<(), PdError> {
+        self.guard_no_fw_update_active()?;
+        match self.tps6699x.execute_hrst(port).await.map_err(|e| self.log_error(e))? {
+            ReturnValue::Success => Ok(()),
+            r => {
+                debug!("Error executing hard reset on port {}: {:#?}", port.0, r);
+                Err(PdError::InvalidResponse)
+            }
+        }
+    }
+
+    async fn get_discovered_svids(&mut self, port: LocalPortId) -> Result<DiscoveredSvids, PdError> {
+        self.guard_no_fw_update_active()?;
+        let svids = self
+            .tps6699x
+            .get_discovered_svids(port)
+            .await
+            .map_err(|e| self.log_error(e))?;
+        debug!("{:?} discovered SVIDs: {:?}", port, svids);
+        let mut sop = Vec::new();
+        for svid in svids.svid_sop().take(sop.capacity()) {
+            let _ = sop.push(svid);
+        }
+
+        let mut sop_prime = Vec::new();
+        for svid in svids.svid_sop_prime().take(sop_prime.capacity()) {
+            let _ = sop_prime.push(svid);
+        }
+
+        Ok(DiscoveredSvids::new(sop, sop_prime))
+    }
+
+    async fn get_discover_identity_sop_response(
+        &mut self,
+        port: LocalPortId,
+    ) -> Result<embedded_usb_pd::vdm::structured::command::discover_identity::sop::ResponseVdos, PdError> {
+        self.guard_no_fw_update_active()?;
+        let data = self
+            .tps6699x
+            .get_received_sop_identity_data(port)
+            .await
+            .map_err(|e| self.log_error(e))?;
+        match data.try_into() {
+            Ok(vdos) => Ok(vdos),
+            Err(e) => {
+                debug!("Error deserializing Received SOP Identity Data: {:?}", e);
+                Err(PdError::Serialize)
+            }
+        }
+    }
+
+    async fn get_discover_identity_sop_prime_response(
+        &mut self,
+        port: LocalPortId,
+    ) -> Result<embedded_usb_pd::vdm::structured::command::discover_identity::sop_prime::ResponseVdos, PdError> {
+        self.guard_no_fw_update_active()?;
+        let data = self
+            .tps6699x
+            .get_received_sop_prime_identity_data(port)
+            .await
+            .map_err(|e| self.log_error(e))?;
+        match data.try_into() {
+            Ok(vdos) => Ok(vdos),
+            Err(e) => {
+                debug!("Error deserializing Received SOP Prime Identity Data: {:?}", e);
+                Err(PdError::Serialize)
+            }
+        }
+    }
 }
 
 impl<M: RawMutex, B: I2c> Retimer for Tps6699x<'_, M, B> {
@@ -887,83 +957,6 @@ impl<M: RawMutex, B: I2c> type_c_interface::controller::max_sink_voltage::MaxSin
             .set_autonegotiate_sink_max_voltage(port, voltage_mv)
             .await
             .map_err(|e| self.log_error(e))
-    }
-}
-
-impl<M: RawMutex, B: I2c> type_c_interface::controller::svid::SvidDiscovery for Tps6699x<'_, M, B> {
-    async fn get_discovered_svids(&mut self, port: LocalPortId) -> Result<DiscoveredSvids, PdError> {
-        self.guard_no_fw_update_active()?;
-        let svids = self
-            .tps6699x
-            .get_discovered_svids(port)
-            .await
-            .map_err(|e| self.log_error(e))?;
-        debug!("{:?} discovered SVIDs: {:?}", port, svids);
-        let mut sop = Vec::new();
-        for svid in svids.svid_sop().take(sop.capacity()) {
-            let _ = sop.push(svid);
-        }
-
-        let mut sop_prime = Vec::new();
-        for svid in svids.svid_sop_prime().take(sop_prime.capacity()) {
-            let _ = sop_prime.push(svid);
-        }
-
-        let svids = DiscoveredSvids::new(sop, sop_prime);
-        Ok(svids)
-    }
-}
-
-impl<M: RawMutex, B: I2c> type_c_interface::controller::hard_reset::HardReset for Tps6699x<'_, M, B> {
-    async fn hard_reset(&mut self, port: LocalPortId) -> Result<(), PdError> {
-        self.guard_no_fw_update_active()?;
-        match self.tps6699x.execute_hrst(port).await.map_err(|e| self.log_error(e))? {
-            ReturnValue::Success => Ok(()),
-            r => {
-                debug!("Error executing hard reset on port {}: {:#?}", port.0, r);
-                Err(PdError::InvalidResponse)
-            }
-        }
-    }
-}
-
-impl<M: RawMutex, B: I2c> type_c_interface::controller::discover_identity::DiscoverIdentity for Tps6699x<'_, M, B> {
-    async fn get_discover_identity_sop_response(
-        &mut self,
-        port: LocalPortId,
-    ) -> Result<embedded_usb_pd::vdm::structured::command::discover_identity::sop::ResponseVdos, PdError> {
-        self.guard_no_fw_update_active()?;
-        let data = self
-            .tps6699x
-            .get_received_sop_identity_data(port)
-            .await
-            .map_err(|e| self.log_error(e))?;
-        match data.try_into() {
-            Ok(vdos) => Ok(vdos),
-            Err(e) => {
-                debug!("Error deserializing Received SOP Identity Data: {:?}", e);
-                Err(PdError::Serialize)
-            }
-        }
-    }
-
-    async fn get_discover_identity_sop_prime_response(
-        &mut self,
-        port: LocalPortId,
-    ) -> Result<embedded_usb_pd::vdm::structured::command::discover_identity::sop_prime::ResponseVdos, PdError> {
-        self.guard_no_fw_update_active()?;
-        let data = self
-            .tps6699x
-            .get_received_sop_prime_identity_data(port)
-            .await
-            .map_err(|e| self.log_error(e))?;
-        match data.try_into() {
-            Ok(vdos) => Ok(vdos),
-            Err(e) => {
-                debug!("Error deserializing Received SOP Prime Identity Data: {:?}", e);
-                Err(PdError::Serialize)
-            }
-        }
     }
 }
